artist contribution formatting changes
these are mostly because the updated artist page will fall back to the
contrib description in Contributors if none is provided for their entry
in Artists; so if someone in Artists were also credited in Contributors
for, say, doing lyrics or a particular instrument, their artist page
would list e.g. "(lyrics)" next to their entry on the track, even though
they were also implicitly responsible for composition (by being in the
Artists list). now it's explicit - this artist will be credited now as
"(composition, lyrics)"!